Make download links configurable
This way we won't have to redeploy once the new links to the windows and the android version are there.

Also this obviously offers more flexibility for providers.
